Updated API to new format to put more in-line with web app functionality
