+++ src/com/warrenrobotics/Tester.java
@@ -1,15 +1,25 @@
 package com.warrenrobotics;
 
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
 public class Tester {
 	public static void main(String[] args) throws Exception{
 		//Time runtime of whole program
 		long curTime = System.currentTimeMillis();
 		//Set event link(s)
-		String eventLink = "https://www.robotevents.com/robot-competitions/vex-robotics-competition/RE-VRC-19-8971.html"; //Random for test
-		//Essentially grab all data and set it(constructor body calls create and write methods)
-		@SuppressWarnings("unused")
+		// future event 2020-03-07 to test 4-week lead time
+		// String defaultLink = "https://www.robotevents.com/robot-competitions/vex-robotics-competition/RE-VRC-19-1124.html";
+		String defaultLink = "https://www.robotevents.com/robot-competitions/vex-robotics-competition/RE-VRC-19-8971.html"; //Random for test
 		//If args isn't empty, use those. Otherwise, default values
-		TeamAPI api = args.length != 0 ? new TeamAPI(args[0]) : new TeamAPI(eventLink);
+		String eventLink = args.length != 0 ? args[0] : defaultLink;
+
+		System.out.printf("%s - Running Program%n", new SimpleDateFormat("MM/dd/yyyy HH:mm:ss").format(new Date()));
+
+		VexEvent event = VexLoader.loadEvent(eventLink, "");
+		SheetsWriter sheetsWriter = new SheetsWriter();
+		sheetsWriter.writeVexEvent(event);
+
 		//Time taken
 		long timeTaken = System.currentTimeMillis() - curTime; 
 		//Print out message
